# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
    "/users/all": {
      "get": {
        "operationId": "UsersController_getAllUsers",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Users retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserResponseDto"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Only admins and moderators can access this resource",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                },
                "examples": {
                  "forbidden": {
                    "summary": "Insufficient permissions",
                    "value": {
                      "success": false,
                      "message": "Only admins and moderators can access this resource",
                      "error": "Forbidden",
                      "statusCode": 403,
                      "timestamp": "2025-02-27T09:05:47.193Z",
                      "path": "/users/all"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all users",
        "tags": [
          "users"
        ]
      }
    },
    "/users/register": {
      "post": {
        "operationId": "UsersController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              },
              "examples": {
                "valid": {
                  "summary": "Valid request",
                  "value": {
                    "email": "user@example.com",
                    "password": "User1@@User1",
                    "firstName": "cse",
                    "lastName": "zag"
                  }
                },
                "invalid": {
                  "summary": "Invalid request",
                  "value": {
                    "email": "invalid-email",
                    "password": "weak",
                    "firstName": "A",
                    "lastName": "B"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User registered successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation errors",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                },
                "examples": {
                  "validationError": {
                    "summary": "Invalid input",
                    "value": {
                      "success": false,
                      "message": "Validation failed",
                      "error": "Bad Request",
                      "statusCode": 400,
                      "timestamp": "2025-02-27T09:05:47.193Z",
                      "path": "/users/register",
                      "errors": {
                        "email": "Invalid email format"
                      }
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Email already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                },
                "examples": {
                  "emailConflict": {
                    "summary": "Email already in use",
                    "value": {
                      "success": false,
                      "message": "Email already exists",
                      "error": "Conflict",
                      "statusCode": 409,
                      "timestamp": "2025-02-27T09:05:47.193Z",
                      "path": "/users/register"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Register a new user",
        "tags": [
          "users"
        ]
      }
    },
    "/users/verify-email": {
      "post": {
        "operationId": "UsersController_verifyEmail",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyEmailDto"
              },
              "examples": {
                "example1": {
                  "value": {
                    "verificationToken": "verification-token"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Email verified successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerifyEmailResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid or expired verification token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                },
                "examples": {
                  "invalidToken": {
                    "summary": "Invalid token",
                    "value": {
                      "success": false,
                      "message": "Invalid or expired verification token",
                      "error": "Bad Request",
                      "statusCode": 400,
                      "timestamp": "2025-02-27T09:05:47.193Z",
                      "path": "/users/verify-email"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Verify user email",
        "tags": [
          "users"
        ]
      }
    },
    "/users/resend-verification": {
      "post": {
        "operationId": "UsersController_resendVerificationEmail",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResendEmailVerificationDto"
              },
              "examples": {
                "example1": {
                  "value": {
                    "email": "user@example.com"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Verification email sent successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResendVerificationResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Email not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                },
                "examples": {
                  "notFound": {
                    "summary": "Email not found",
                    "value": {
                      "success": false,
                      "message": "Email not found",
                      "error": "Not Found",
                      "statusCode": 404,
                      "timestamp": "2025-02-27T09:05:47.193Z",
                      "path": "/users/resend-verification"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Resend verification email",
        "tags": [
          "users"
        ]
      }
    },
    "/users/refresh-token": {
      "post": {
        "operationId": "UsersController_refreshToken",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              },
              "examples": {
                "example1": {
                  "value": {
                    "refreshToken": "refresh-token"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token refreshed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefreshTokenResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Invalid or expired refresh token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                },
                "examples": {
                  "unauthorized": {
                    "summary": "Invalid token",
                    "value": {
                      "success": false,
                      "message": "Invalid or expired refresh token",
                      "error": "Unauthorized",
                      "statusCode": 401,
                      "timestamp": "2025-02-27T09:05:47.193Z",
                      "path": "/users/refresh-token"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Refresh JWT token",
        "tags": [
          "users"
        ]
      }
    },
    "/users/login": {
      "post": {
        "operationId": "UsersController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginUserDto"
              },
              "examples": {
                "example1": {
                  "value": {
                    "email": "user@example.com",
                    "password": "Password123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User logged in successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                },
                "examples": {
                  "unauthorized": {
                    "summary": "Invalid credentials",
                    "value": {
                      "success": false,
                      "message": "Invalid credentials",
                      "error": "Unauthorized",
                      "statusCode": 401,
                      "timestamp": "2025-02-27T09:05:47.193Z",
                      "path": "/users/login"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Login user",
        "tags": [
          "users"
        ]
      }
    },
    "/users/change-password": {
      "put": {
        "operationId": "UsersController_changePassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              },
              "examples": {
                "example1": {
                  "value": {
                    "oldPassword": "OldPassword123",
                    "newPassword": "NewPassword123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password changed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangePasswordResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Invalid old password",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                },
                "examples": {
                  "unauthorized": {
                    "summary": "Invalid old password",
                    "value": {
                      "success": false,
                      "message": "Invalid old password",
                      "error": "Unauthorized",
                      "statusCode": 401,
                      "timestamp": "2025-02-27T09:05:47.193Z",
                      "path": "/users/change-password"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Change user password",
        "tags": [
          "users"
        ]
      }
    },
...
  },
  "info": {
    "title": "Smart Airport API",
    "description": "API documentation for the Smart Airport application",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "UserResponseDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "example": "cse"
          },
          "lastName": {
            "type": "string",
            "example": "zag"
          },
          "email": {
            "type": "string",
            "example": "user@example.com"
          },
          "roles": {
            "example": [
              "user"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "isVerified": {
            "type": "boolean",
            "example": false
          }
        },
        "required": [
          "firstName",
          "lastName",
          "email",
          "roles",
          "isVerified"
        ]
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {}
      },
      "RegisterResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object",
            "properties": {
              "message": {
                "type": "string",
                "example": "User registered successfully"
              },
              "user": {}
            }
          }
        },
        "required": [
          "success",
          "data"
        ]
      },
      "VerifyEmailDto": {
        "type": "object",
        "properties": {
          "verificationToken": {
            "type": "string",
            "example": "e1dASL2qpwd2ld!@asca",
            "description": "Verification token"
          }
        },
        "required": [
          "verificationToken"
        ]
      },
      "VerifyEmailResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "Email verified successfully"
          }
        },
        "required": [
          "success",
          "message"
        ]
      },
...
    }
  }
}